requirements.txt
# Video Processing
opencv-python>=4.9.0.80
numpy>=1.26.0
pandas>=2.2.0

# Pose Estimation
mediapipe>=0.10.14

# Object Detection (YOLOv8)
ultralytics>=8.2.0

# Video Downloading
yt-dlp>=2024.4.9

# Config Management
pyyaml>=6.0.1
python-dotenv>=1.0.1

# Logging & Utilities
tqdm>=4.66.1
matplotlib>=3.8.2

# Optional GPU Support (if you have NVIDIA GPU)
# torch>=2.2.0
# torchvision>=0.17.0



#    pip install -r requirements.txt